Added verifier
Necessary to build a sufficiently random and secure verification code for OAuth API requests that adheres to the PKCE standard. Due to the nature of the Base64 URL encoding, the character range is smaller than the PKCE spec because the randomized string will end up larger than the length given. This means that the 32..96 range equates to 43..128 which still adheres to the [specification](https://datatracker.ietf.org/doc/html/rfc7636#section-4.1).